Refactor now using only /proc/<pid>
This is because we can't match a process by the local port, as the
packet we capture has a different port (has been natted) when the
containers are running in something more complex than a basic docker
network.

That also means that we have to re-scan the whole set of processes on
every trigger.

Signed-off-by: David Caro <dcaro@wikimedia.org>